fix(backend): add explicit type annotations to resolve TS2304 errors
- Add ProductModel type alias for Product in ProductService.ts
- Move ProductListOptions interface from ProductService to types/index.ts
- Import ProductListOptions in ProductService from types
- Eliminates all 13 TS2304 'Cannot find name' errors
- TS7006 errors already resolved by Phase A/B codemod work
