+++ backend/src/services/ProductService.ts
@@ -15,31 +15,16 @@
 import { Op, Transaction } from 'sequelize';
 import { sequelize, Product, Category, InventoryMovement, User } from '../models/index.js';
 import { AppError } from '../middleware/error.middleware.js';
-import type { ProductAttributes, ProductType, GenericProductAttributes } from '../types/index.js';
+import type {
+  ProductAttributes,
+  ProductType,
+  GenericProductAttributes,
+  ProductListOptions,
+} from '../types/index.js';
 import type { InventoryMovementCreation } from '../models/InventoryMovement.js';
 
-export interface ProductListOptions {
-  page?: number;
-  limit?: number;
-  platform?:
-    | 'netflix'
-    | 'disney_plus'
-    | 'spotify'
-    | 'hbo_max'
-    | 'amazon_prime'
-    | 'youtube_premium'
-    | 'apple_tv'
-    | 'other';
-  isActive?: boolean; // true for active, false for inactive, undefined for both
-  minPrice?: number;
-  maxPrice?: number;
-  // Extended filters for generic products
-  type?: ProductType;
-  categoryId?: string;
-  minStock?: number;
-  maxStock?: number;
-  search?: string;
-}
+/** Sequelize Product model instance type */
+type ProductModel = Product;
 
 export interface PaginatedProducts {
   rows: ProductModel[];
